chore(test): move FakeBitboxCredentials out of lib/ into test/helper/
The fake holds a hardcoded test private key and only exists to drive
unit/integration tests. It does not belong in the production tree even
though tree-shaking strips it from release builds today. Moving it to
test/helper/ alongside pump_app keeps test-only state out of lib/ and
makes ad-hoc 'find references' searches in the production code less
noisy.

The accompanying fake_bitbox_credentials_test.dart moves with it so the
test sits next to its subject. Five call sites switch to the new
relative import path.
